+++ test/e2e/tests/metrics/state-snapshots/errors-before-init-opt-in-background-state.json
@@ -129,10 +129,12 @@
       "openSeaEnabled": false,
       "overrideContentSecurityPolicyHeader": true,
       "preferences": {
+        "defaultAddressScope": "string",
         "hideZeroBalanceTokens": false,
         "petnamesEnabled": "boolean",
         "shouldShowAggregatedBalancePopover": "boolean",
         "showConfirmationAdvancedDetails": false,
+        "showDefaultAddress": "boolean",
         "showExtensionInFullSizeView": false,
         "showFiatInTestnets": false,
         "showMultiRpcModal": "boolean",
