+++ ui/components/multichain-accounts/multichain-account-network-group-with-default-address/index.ts
@@ -1,0 +1,2 @@
+export { MultichainAccountNetworkGroupWithDefaultAddress } from './multichain-account-network-group-with-default-address';
+export type { MultichainAccountNetworkGroupWithDefaultAddressProps } from './multichain-account-network-group-with-default-address';
